small fixes

+++ yaskoam.mrz2.lab1/src/main/java/yaskoam/mrz2/lab1/neuro/NeuralNetwork.java
@@ -2,7 +2,7 @@ package yaskoam.mrz2.lab1.neuro;
 
 import java.util.Calendar;
 
-import org.jblas.FloatMatrix;
+import org.jblas.DoubleMatrix;
 /**
  * @author Q-YAA
  */
@@ -12,61 +12,61 @@ public class NeuralNetwork {
 
     private int p;
 
-    private float maxError;
+    private double maxError;
 
-    private float a;
+    private double a;
 
-    private FloatMatrix W1;
+    private DoubleMatrix W1;
 
-    private FloatMatrix W2;
+    private DoubleMatrix W2;
 
-    public NeuralNetwork(int n, int p, float maxError, float a) {
+    public NeuralNetwork(int n, int p, double maxError, double a) {
         this.n = n;
         this.p = p;
         this.maxError = maxError;
         this.a = a;
 
-        W1 = new FloatMatrix(createRandomArray(n, p));
-        W2 = new FloatMatrix(createRandomArray(p, n));
+        W1 = new DoubleMatrix(createRandomArray(n, p));
+        W2 = new DoubleMatrix(createRandomArray(p, n));
     }
 
 
-    public float[][] compress(float[][] segments) {
-        float[][] compressed = new float[segments.length][p];
+    public double[][] compress(double[][] segments) {
+        double[][] compressed = new double[segments.length][p];
 
         for (int i = 0; i < segments.length; i++) {
-            FloatMatrix X = new FloatMatrix(new float[][]{segments[i]});
+            DoubleMatrix X = new DoubleMatrix(new double[][]{segments[i]});
             compressed[i] = X.mmul(W1).data;
         }
 
         return compressed;
     }
 
-    public float[][] decompress(float[][] segments) {
-        float[][] decompressed = new float[segments.length][n];
+    public double[][] decompress(double[][] segments) {
+        double[][] decompressed = new double[segments.length][n];
 
         for (int i = 0; i < segments.length; i++) {
-            FloatMatrix Y = new FloatMatrix(new float[][]{segments[i]});
+            DoubleMatrix Y = new DoubleMatrix(new double[][]{segments[i]});
             decompressed[i] = Y.mmul(W2).data;
         }
 
         return decompressed;
     }
 
-    public void learn(float[][] segments) {
+    public void learn(double[][] segments) {
 
-        float error;
+        double error;
 
         do {
 
             long currentTime = Calendar.getInstance().getTimeInMillis();
 
             // learn
-            for (float[] segment : segments) {
-                FloatMatrix X = new FloatMatrix(new float[][]{segment});
+            for (double[] segment : segments) {
+                DoubleMatrix X = new DoubleMatrix(new double[][]{segment});
 
-                FloatMatrix Y = X.mmul(W1);
-                FloatMatrix deltaX = Y.mmul(W2).sub(X);
+                DoubleMatrix Y = X.mmul(W1);
+                DoubleMatrix deltaX = Y.mmul(W2).sub(X);
 
                 W1 = W1.sub((X.transpose()).mmul(deltaX).mmul((W2).transpose()).mmul(a));
                 W2 = W2.sub(Y.transpose().mmul(deltaX).mmul(a));
@@ -75,10 +75,10 @@ public class NeuralNetwork {
             error = 0;
 
             // calculate error
-            for (float[] segment : segments) {
-                FloatMatrix X = new FloatMatrix(new float[][]{segment});
-                FloatMatrix Y = X.mmul(W1);
-                FloatMatrix deltaX = Y.mmul(W2).sub(X);
+            for (double[] segment : segments) {
+                DoubleMatrix X = new DoubleMatrix(new double[][]{segment});
+                DoubleMatrix Y = X.mmul(W1);
+                DoubleMatrix deltaX = Y.mmul(W2).sub(X);
 
                 for (int j = 0; j < deltaX.length; j++) {
                     error += deltaX.get(0, j) * deltaX.get(0, j);
@@ -91,12 +91,12 @@ public class NeuralNetwork {
         while (error > maxError);
     }
 
-    private float[][] createRandomArray(int n, int m) {
-        float[][] array = new float[n][m];
+    private double[][] createRandomArray(int n, int m) {
+        double[][] array = new double[n][m];
 
         for (int i = 0; i < n; i++) {
             for (int j = 0; j < m; j++) {
-                array[i][j] = (float) Math.random();
+                array[i][j] = Math.random();
             }
         }
