+++ yaskoam.mrz2.lab1/src/main/java/yaskoam/mrz2/lab1/neuro/NeuroPixel.java
@@ -5,32 +5,32 @@ package yaskoam.mrz2.lab1.neuro;
  */
 public class NeuroPixel {
 
-    private float r;
+    private double r;
 
-    private float g;
+    private double g;
 
-    private float b;
+    private double b;
 
-    public NeuroPixel(float r, float g, float b) {
+    public NeuroPixel(double r, double g, double b) {
         this.r = r;
         this.g = g;
         this.b = b;
     }
 
-    public float getR() {
+    public double getR() {
         return r;
     }
 
-    public float getG() {
+    public double getG() {
         return g;
     }
 
-    public float getB() {
+    public double getB() {
         return b;
     }
 
-    public static float[] convertToSimple(NeuroPixel[] array) {
-        float[] simpleArray = new float[array.length * 3];
+    public static double[] convertToSimple(NeuroPixel[] array) {
+        double[] simpleArray = new double[array.length * 3];
 
         int index = 0;
         for (NeuroPixel pixel : array) {
